requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scipy
scikit-learn
plotly
google-generativeai
requests
